% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{birthweight}
\alias{birthweight}
\title{Birthweight Dataset}
\format{
A haven-imported dataframe with 189 rows and 10 variables.
}
\source{
MASS Package
}
\usage{
birthweight
}
\description{
Dataset for MGH IHP Statistics Course RS932 from the MASS package
}
\details{
Data from a study examining factors that affect the birth weights of infants.
The researchers were interested in a variety of demographic, health, and pregnancy-related variables
that might impact birth weight. The dataset is included in the MASS package and
has data on 189 births at the Baystate Medical Centre, Springfield, Massachusetts
during 1986 (Venables & Ripley, 2002).
\itemize{
\item \strong{pt_code:}	Patient Code
\item \strong{moms_age:}	Mother's age
\item \strong{moms_nl_wt:}	Mother's pre-pregnancy weight
\item \strong{moms_race:}	Mother's race (1 = white, 2 = African-American, 3 = Asian/PI)
\item \strong{moms_smoking:} Mother's smoking status during pregnancy (0=no, 1 = yes)
\item \strong{prev_premlabor:} Number of previous pre-term labors
\item \strong{hypertension:} Mother's history of hypertension (0=no, 1 = yes)
\item \strong{Uterine_irrit:} presence of uterine irritability (0=no, 1 = yes)
\item \strong{MD_visits:} # of MD visits in 1st trimester
\item \strong{birthweight:} birthweight in grams
}
}
\keyword{datasets}
