% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Dataset}
\format{
A haven-imported dataframe with 403 rows and 22 variables.
}
\source{
http://staff.pubhealth.ku.dk/~tag/Teaching/share/data/Diabetes.html
}
\usage{
diabetes
}
\description{
Dataset for MGH IHP Statistics Course RS932
}
\details{
These data are courtesy of Dr John Schorling, Department of Medicine, University of Virginia School of Medicine.
The dataset contains data from 403 participants in a study to understand the prevalence of obesity, diabetes, and
other cardiovascular risk factors in central Virginia for African Americans.
\itemize{
\item \strong{id:}	ID
\item \strong{total_cholesterol:}	Total Cholesterol
\item \strong{stabilized.glucose:}
\item \strong{hdl:}
\item \strong{total_chol_hdl_ratio:}
\item \strong{glycosoloatedhgb:}
\item \strong{location:} Buckingham or Louisa
\item \strong{age:} Age
\item \strong{gender:} gender (male = 0, female = 1)
\item \strong{height:} height in inches
\item \strong{weight:} height in pounds
\item \strong{bp.1s:} systolic blood pressure
\item \strong{bp.1d:} diastolic blood pressure
\item \strong{waist:} waist circumference
\item \strong{hip:} hip circumference
\item \strong{diabetes_diagnosis:}  Diabetes diagnosis (0 = no, 1 = yes)
\item \strong{BMI:}
\item \strong{waist_hip_ratio:}
\item \strong{BMI_cat_5:} Categorical BMI variable (1 = not obese, 2 = overweight, 3 = Obese Class 1, 4 = Obese Class 2, 5 = Obese Class 3)
\item \strong{age_cat:}  Categorical age variable (1 = <35, 2 = 35-49, 3 = 50-64, 4 = >=64) )
\item \strong{waist_hip_cat:} Categorical waist to hip ratio (0 = normal, 1 = abnormally high)
\item \strong{BMI_cat_3:}  Categorical BMI variable (1 = not overweight/obese, 2 = overweight, 3 = obese)
}
}
\keyword{datasets}
