% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{neuro_data}
\alias{neuro_data}
\title{Neuro Dataset}
\format{
A dataframe with 4878 rows and 64 variables.
}
\source{
RS932
}
\usage{
neuro_data
}
\description{
Dataset for MGH IHP Statistics Course RS932
}
\details{
Data collected from a large healthcare system with several rehabilitation clinics.
\itemize{
\item \strong{gender:}	gender
\item \strong{age:}	age
\item \strong{ethnic:}	ethnicity
\item \strong{dompart:}	has domestic partner
\item \strong{onsett:}	living setting prior to onset
\item \strong{complex:}	medical complexity score
\item \strong{cyclelos:}	total therapy days
\item **episod:**e	total length of stay
\item \strong{impgroup:}	impairment group
\item \strong{admfim:}	total FIM within 72 hours admission
\item \strong{disfim:}	total FIM at discharge
\item \strong{fimgain:}	FIM change
\item \strong{loseff:}	FIM gain/Rx adm date to Rx dc date
\item \strong{dcsett:}	living setting after discharge
\item \strong{thours:}	total number of hours of therapy
\item \strong{tvisits:}	number of therapy visits
\item \strong{ptmins:}	total PT minutes
\item \strong{ptvisits:}	total PT visits
\item \strong{otmins:}	total OT minutes
\item \strong{otvisits:}	total OT visits
\item \strong{stmins:}	total ST minutes
\item \strong{stvisits:}	total ST visits
\item \strong{admmotor:}	FIM motor at admission
\item \strong{admcog:}	FIM cognitive at admission
\item \strong{dcmotor:}	FIM motor at discharge
\item \strong{dccog:}	FIM cognitive at discharge
\item \strong{aeating:}	admission eating
\item \strong{agroomin:}	admission grooming
\item \strong{abath:}	admission bathing
\item \strong{adressup:}	admission dsg upper
\item \strong{adresslo:}	admission dsg lower
\item \strong{atoil:}	admission toilet
\item \strong{abladder:}	admission bladder
\item \strong{abowel:}	admission bowel
\item \strong{atranbcw:}	admission chair transfer
\item \strong{atrantoi:}	admission toilet transfer
\item \strong{atrantub:}	admission tub transfer
\item \strong{awalkwc:}	admission walking or wc mobility
\item \strong{astairs:}	admission stairs
\item \strong{acomp:}	admission comprehension
\item \strong{aexpress:}	admission expression
\item \strong{asocint:}	admission social interaction
\item \strong{aprob:}	admission problem solving
\item \strong{amemory:}	admission memory
\item \strong{deating:}	discharge eating
\item \strong{dgroomin:}	discharge grooming
\item \strong{dbath:}	discharge bathing
\item \strong{ddressup:}	discharge dsg upper
\item \strong{ddresslo:}	discharge dsg lower
\item \strong{dtoil:}	discharge toilet
\item \strong{dbladder:}	discharge bladder
\item \strong{dbowel:}	discharge bowel
\item \strong{dtranbcw:}	discharge chair transfer
\item \strong{dtrantoi:}	discharge toilet transfer
\item \strong{dtrantub:}	discharge tub transfer
\item \strong{dwalkwc:}	discharge walking or wc mobility
\item \strong{dstairs:}	discharge stairs
\item \strong{dcomp:}	discharge comprehension
\item \strong{dexpress:}	discharge expression
\item \strong{dsocint:}	discharge social interaction
\item \strong{dprob:}	discharge problem solving
\item \strong{dmemory:} discharge memory
\item \strong{age_cat:}	age categories
}
}
\keyword{datasets}
