% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{neurological_FIM}
\alias{neurological_FIM}
\title{Neurological FIM Dataset}
\format{
A dataframe with 721 rows and 24 variables.
}
\source{
RS 932
}
\usage{
neurological_FIM
}
\description{
Dataset for MGH IHP Statistics Course RS932
}
\details{
Data collected from a clinical trial at a large healthcare system with several rehabilitation clinics.
Patients receive one of three different modes of treatment delivery:  group and individual therapy per day,
individual therapy twice a day, or group therapy twice a day. Primary outcome of interest is FIM discharge scores.
\itemize{
\item \strong{PtID:}	Patient ID
\item \strong{facnum:}	Facility Code #
\item \strong{age:}	age
\item \strong{ethnic:}	ethnicity
\item \strong{admfim:}	total FIM within 72 hours admission
\item \strong{disfim:}	total FIM at discharge
\item \strong{treatment:}  treatment group (1 = group + individual therapy each day,
2 = individual therapy 2x/day, 3 = group therapy 2x/day) )
\item \strong{cyclelos:}	total therapy days
\item \strong{episode:} 	total length of stay
\item \strong{dccomm:}  Discharge to community (0=no, 100 =yes)
\item \strong{thours:}	total number of hours of therapy
\item \strong{total_sessions:}	number of therapy visits
\item \strong{ptmins:}	total PT minutes
\item \strong{ptsessions:}	total PT visits
\item \strong{otmins:}	total OT minutes
\item \strong{otsessions:}	total OT visits
\item \strong{stmins:}	total ST minutes
\item \strong{stsessions:}	total ST visits
\item \strong{admmotor:}	FIM motor at admission
\item \strong{admcog:}	FIM cognitive at admission
\item \strong{dcmotor:}	FIM motor at discharge
\item \strong{dccog:}	FIM cognitive at discharge
\item \strong{gender:} gender (1 = male, 2 = female)
\item \strong{FIM_change:}  FIM change score from admission to discharge
}
}
\keyword{datasets}
