% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{private_practice}
\alias{private_practice}
\title{Private Practice Dataset}
\format{
A dataframe with 262 rows and 7 variables.
}
\source{
Pollack, J., VanEpps, E. M., & Hayes, A. F. (2012). The moderating role of social ties on entrepreneurs'
depressed affect and withdrawal intentions in response to economic stress.
\emph{Journal of Organizational Behavior, 33,} 789-810.
}
\usage{
private_practice
}
\description{
Dataset for MGH IHP Statistics Course RS932 based on Pollack et al. (2012) described in Hayes (2017)
}
\details{
This study examined whether economic stress leads to a desire to disengage from private practice as a result of
the pessimism that the stress creates.  This dataset includes data from a networking group of owners of private
rehabilitation practices. They were interviewed to determine their perceptions of how their practices were
doing, and their emotional and cognitive reactions to the current healthcare climate. The responses were
used to calculate a 1) measure of economic distress (a mean of 2 items rated on a scale of 1 to 7, the higher
the score, the greater the distress); 2) a measure of how optimistic or pessimistic they felt about the future
of their practice, (a mean of 6 items rated on a scale of 1 to 5, with higher scores meaning more pessimistic);
3) their intentions to withdraw from private practice in the near future (a mean of 3 items that were rated
on a scale of 1 to 7 with higher intention to withdraw).
\itemize{
\item \strong{own_yrs:}	years owning private practice
\item \strong{Ecstress:}	economic stress (higher scores = more stress)
\item \strong{affect:}	Pessimism (higher scores = more pessimism)
\item \strong{disengag:}	Withdrawal intentions (higher more likely to withdraw)
\item \strong{sex:} gender (1 = male, 0 = female)
\item \strong{age:} age
\item \strong{self_eff:} entrepreneurial self-efficacy (higher = higher self-efficacy)
}
}
\keyword{datasets}
