% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{therapist}
\alias{therapist}
\title{Therapist Self-Confidence Dataset}
\format{
A dataframe with 164 rows and 5 variables.
}
\source{
RS932
}
\usage{
therapist
}
\description{
Dataset for MGH IHP Statistics Course RS932
}
\details{
This is a study of 164 newly graduated PTs, OTs, and SLPs. The participants were asked to rate their self-confidence
managing each of their clients/patients over the period of 1 year.  The scale was a visual analog scale with 1 as
completely lacking in confidence and 10 as totally confident.  An average score across all clients/patients was
compiled for each therapist. They were also asked to rate their perception of the support they received from
their direct supervisor over the 1 year period.  The scale was 0-100\%, with 100\% being full support.
The therapists’ client/patients each rated their level of satisfaction with their care on a scale of -1,
totally unsatisfied, to +1, completely satisfied.  An average score across all clients/patients was computed
for each therapist.
\itemize{
\item \strong{caseID:}	Case ID number
\item \strong{pt_satis:}	Average patient satisfaction score (-1 to 1)
\item \strong{gender:}	gender (0 = female, 1 = male)
\item \strong{support:}	Supervisor Support (0-100\%)
\item \strong{selfconf:}	Average self-confidence rating (1-10)
}
}
\keyword{datasets}
