% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{tvability}
\alias{tvability}
\title{TV Ability Data}
\format{
A data frame with 500 rows and 5 variables
}
\source{
\url{https://tzkeith.com/data-files/}
}
\usage{
tvability
}
\description{
Data from the NELS dataset used in the Keith textbook, \emph{Multiple Regression and Beyond, 3rd Edition}.
}
\details{
Data pulled from the NELS dataset that appears in Chapter 8 of Keith (starting on page 162).
The data come from a study examining the relationship between TV viewing and ability in their
effects on achievement. Previous research has suggested a negative relationship between
TV viewing and achievement; however, that relationship might vary depending on a child's abilities.
\itemize{
\item \strong{SES:}	Family SES level, z-scored
\item \strong{ABILITY:}	A composite of 6 different measures of verbal and non-verbal abilities
\item \strong{TV:} Average time per day spent watching TV
\item \strong{ACHIEVE:}	T score composite of reading and math
\item \strong{ABIL_3:} Ability variable categorized into 3 levels (low, med, high)
}
}
\keyword{datasets}
